pyproject.toml
[project]
name = "my-digital-being"
version = "0.1.0"
description = "A sophisticated Python framework for creating intelligent digital entities"
requires-python = ">=3.11"
dependencies = [
    "composio-core>=0.6.7",
    "composio-openai==0.6.7",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "replit>=4.1.0",
    "requests-oauthlib>=2.0.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.1",
    "websockets==12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["my_digital_being*"]
